Reads the DESTATIS GV100 Format

The GV100 format is used by DESTATIS to publish the German municipality register

read_gv100(file, stzrt, version = NULL, encoding = "iso-8859-1", ...)

Arguments

file

path to file

stzrt

integer to select the administrative level (see details)

version

which GV100 version. If NULL the version is guessed based on the file name.

encoding

encoding of the file

...

other parameters passed to read_fwf

Value

a data.frame.

Details

The Gemeindeverzeichnis (municipality register) is published in a fixed width file refered to as "GV1000 ASCII Format" by DESTATIS. The register features the list of municipality and higher order administrative units. The function is a wrapper around [readr::read_fwf()].

There are two types of files: One feature the administrative information (version="AD") and one with non-administrative (version="NAD"). If version=NULL, read_gv100() guess the type based on the file name.

To select a particular administrative unit use the stzrt argument (Satzart). For the AD version, the following choices are possible:

10 - Länder (states) 20 - Regierungsbezirke 30 - Regionsdaten (only Baden-Württemberg) 40 - Kreise (counties) 50 - Gemeindeverbandsdaten 60 - Gemeinden (municipalities)

For the NAD version only:

41 - Kreise (counties) 61 - Gemeinden (municipalities)

See also

Examples

 if (FALSE) {
  
    d <- read_gv100("GV100NAD31122016.asc", stzrt=60)

  }